ignore .git

+++ setup.sh
@@ -6,7 +6,7 @@ dotfiles=`ls -A`
 cd ~
 for dotfile in $dotfiles
 do
-  if [ "setup.sh" != "$dotfile" ]; then
+  if [ "setup.sh" != "$dotfile" -a ".git" != "$dotfile" ]; then
     rm -f "./$dotfile"
     ln -s "$current/$dotfile" "./$dotfile"
   fi
